requirements.txt
# Core Data Science
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
scikit-learn>=1.3.0

# Graph Machine Learning
torch>=2.0.0
torch-geometric>=2.4.0
networkx>=3.1

# Knowledge Graph Utilities
pykeen>=1.10.0

# Cheminformatics (optional molecular features)
rdkit>=2023.3.1

# RAG and LLM
chromadb>=0.4.0
sentence-transformers>=2.2.0
openai>=1.0.0
langchain>=0.1.0

# MLOps
mlflow>=2.8.0

# Web Interface
streamlit>=1.28.0
plotly>=5.17.0

# Utilities
pyyaml>=6.0
tqdm>=4.66.0
requests>=2.31.0
python-dotenv>=1.0.0

# Biomedical Data Access
biopython>=1.81  # For PubMed/NCBI Entrez API
